refactor(how-why): factor out issueDepDot to issueDiagram.js
